+++ packages/vitepress/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @to-skills/vitepress
 
+## 0.2.21
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @to-skills/core@1.1.2
+
 ## 0.2.20
 
 ### Patch Changes